Fixed issue with sometimes not finding arrays
The function to find the end of each array was designed badly. It required a blast hit that was not part of the array to follow all the array repeats in order to know when the array was complete and that it should store the finished array. Now it stores the finished array either when it sees a blast hit outside the array or if this is the last repeat in the blast results list.