Major change to output image scaling.
Instead of setting the image size at 10X10 inches and trying to scale everything to fit in that, now the image size changes and the objects in the image remain the same size. Scaling everything to fit in a 10X10 never worked properly as text size scales differently and some things ended up looking stretched. This looks much better over different tree shapes.